Bugfix: Error in customer profile page when gender is not set
When a registered customer logs in for the first time, his/her data is incomplete yet, since registration form ask him/her only an e-mail and password. In profile form, any other customer information (such as full name, birthdate and others) appears as blank.

The problem was that in the gender field, profile page was assuming that customer's gender is already set, even if this customer hadn't set a gender yet.

This commit solves the issue with a conditional, so gender is shown only if it's already set, and shows as blank if not.