Avoid page scroll events when cart sidebar is open
In mobile devices, a "pop-jump" effect appears when cart sidebar is open and user scrolls down in the page, causing that sidebar height is not 100% and leaving a gap for a few seconds (before sidebar height is automatically adjusted).

This commit solves this glitch by adding an "overflow-y: none" to page body, so scroll up/down is not allowed when sidebar is open (if sidebar is closed), page scroll works as usual).