Hide the native search/cancel button in search box
In Chromium-based web browsers, a search/cancel button (with a blue "X") appears when a search input gets focus. Since Firefox doesn't show any button, a custom search/cancel button was implemented for all browsers.

This commit add a CSS rule that hides the native button, so it won't appear in any browser.